Fix missing URL in detail view of own quote tweets
Even though we ask for "extended" tweet mode, Twitter doesn't give
EXACTLY the same response to posting a tweet as when the tweet
comes back in the timeline.

When we tweet, the quoted tweet is referenced BUT its URL isn't in
the tweet text (and hence it isn't in the URL entities list). This
means that viewing your own tweet (e.g. as a quick way to go back
to the quoted tweet) wasn't working.

When the same tweet comes in the timeline then the URL *is*
included.

We now check the URL entities of any injected tweet for the quoted
tweet's URL and put it back in the text if it isn't there.