Change links to RTs so they work better in New Twitter
This makes the name and ID consistent, and uses the RTed details.
Clicking "Source" now goes to the RTed tweet, not to the RT (which
New Twitter shows as a tweet of the RTer as "RT @<RTed-user>: …"),
which appears to have no link to the RTed tweet itself.

The down-side is that you can't link to the RTing user's RT. The
up-side is that you can see the original tweet and its replies
(which seems like a MUCH more likely goal when Twitter API makes
threading so hard)