Re #9: Record and re-load images
Baedert added "draft" texts, but now we support draft images as well.

If a TLS error happens and the user doesn't close the compose window
then we can use the existing IDs because they're live for ~24h and the
user probably won't wait that long.

If they close the window then it could be re-opened the next day so we
can't rely on the image IDs on the Twitter server still being live, so
we need to re-upload them.