Closes #51: Fix segfaults with function definition
The Open Build Server was building PIE builds. These crashed.
Local dev builds were non-PIE and didn't crash.

Apparently, Vala/C will happily compile and let you reference a function
that wasn't defined in the header. And then it'll crash when it's
position independent (presumably because it can't guarantee where
the function will be in the same way).

By defining the get_quoted_medias function in the header file,
we fix the crash.