Fixes #18 - Mark protected accounts with padlock
Using the Adwaita "readonly" emblem (under LGPL/CC-BY-SA3)
because it works well over any avatar.

Included marker in completion box as well.

Not included in DMs because a) it's not directly relevant at that point
and b) Twitter doesn't put that info in the JSON (we currently
read it from the database and assume it won't change, which
protected status is more likely to do)

Also added artist credits to About to reference license