Give better feedback replying to deleted tweets
It still isn't perfect, but "385: You attempted to reply to a Tweet that is deleted or not visible to you. (ComposeTweetWindow.vala:XXX)"
is still clearer than "Forbidden", even if it includes some info that
will confuse some users.

Without that info, there's no reliable way to find the location for
real errors (unless we can build a dialog with a "Show detail" expander)

Fixes #8