Re #59: Set max size based on current monitor
This depends on GDK 3.22. The old "screen" functions were deprecated.

This should solve sizing, but won't do anything about positioning (which is
controlled by the window manager). Media may still sit across monitors if it
is bigger than the distance between the window and the edge of the
monitor, IF that edge is one that it can overflow across.