Fix UserCounter tests
Apparently the tests were picking up the ACCOUNTS_SQL_VERSION
from the auto-generated cawbird.h. But that normally gets built in
the build directory, so we don't need it.

I think. But this is C and Vala, so more stuff may break later. But
tests and builds are working for now. So we're probably okay.