Re #164: Handle smooth scroll events
We checked delta_y to see if we were scrolling up.
But the docs[1] say:

Some GDK backends can also generate “smooth” scroll events, which can
be recognized by the gdk_scroll_smooth scroll direction. For these,
the scroll deltas can be obtained with get_scroll_deltas.

We now use the function to get the deltas to handle normal and smooth.

[1] https://valadoc.org/gdk-3.0/Gdk.EventScroll.html