Detach signal handler when receiver is finalized
Previously, it was possible for the hires_progress callback to be called
after the associated CbMediaImageWidget had already been finalized,
since the handler is only automatically disconnected when the emitter
(the CbMedia) is finalized. Track the handler ID and also the CbMedia
inside the CbMediaImageWidget, so that the handler can be disconnected
in the finalizer if it had been connected before.

Fixes #196.