Fix DM load order
DMs would often load out of sequence. This was because calling async
methods in a loop does not guarantee execution order. This caused
the DM thread to jump around.

Also, we used to scroll to the end before all the asyncs had finished,
which isn't useful when more items then load in.