Allow scrolling back to older DMs
This only loads older DMs from the database. Loading older DMs
from the API is separate (#3)

Works best with scroll wheel/page up. Dragging the scrollbar
triggers a chain of loads because loading moves where the scrollbar
is within the trough and then the UI tries to move it to the mouse again.