Improved solution for #164
Use the pre-existing "don't scroll" functionality and a maximised
container to ensure that replies load above the current tweet
and the scrollbar moves down so that we're not doing weird
things with cropping widgets to get scroll-back for replies.

This should solve usability, discoverability *and* mobile issues.

It could be improved when moving around within a thread, but the
benefits outweight the drawbacks.