Re #270: Do not call a test "shutdown"
Based on some debugging and the ordering of events in the log,
"shutdown" appears to be a special function name and if it gets
called before the tests are run (which is non-determinate because of
the async calls) then it shuts down the downloader and breaks things.

After renaming the tests it ran perfectly for a dozen attempts.

Fixes #270