Use Account's proxy in UserStream
This allows consumer keys to be more easily swapped out, and stops us
creating a new REST proxy within the UserStream when we have a REST
proxy for the user.

Also removed "Stress test" setting because a) it isn't used now,
b) it became the same as non-stress test and c) we never used it anyway!