Re #296: Add "default 1px size" check to height calculation
We already did this in the allocation but not the height.
This didn't seem to cause problems in timelines for me, but was
visible on the Settings tweet, which might be mirroring the
behaviour in the bug.

Also, initial "responsive" testing showed a default size of 1, but
during debugging this it occasionally shows 0. Both are "not properly
allocated" sizes.

Fixes #296