Re #244: Add "reload" button if image load fails
This uses a funky widget approach of having a bin with a button when
needed but normally painting over it with the image.

The button connects to a reload function that resets and redownloads
the thumbnail.

Only tested with forced errors.

The translations have picked up "Could not load image" which is not
right but at least informative as a guess and might get users to click
the button