Re #333: Switch to secrets as build options
Apparently Meson can't do anything useful to create files prior to build
so we'll use Meson options and config values.

This has the advantage that we're not reading resources. But it has the
disadvantage that dev builds can't keep consumer_{k,s}.dat locally
as files that are less likely to get lost than build options.