Re #391: Sort DMs entities before text transform
If there are a combination of URLs and mentions and hashtags then
they won't be in order. We already sorted them in CbTypes.c for tweets
but not here for DMs.

We use a bubble sort because even DMs aren't likely to have many
text entities.

DMManager isn't affected because it only expands URLs (so they're not
showing t.co links in notifications)

Fixes #391