Dirs: create user config dir if it doesn't exist
Cawbird assumes the user config dir (usually `~/.config`) always exists,
which might not be the case, especialy for sandboxed or other
container-based environments. This leads to an error when trying to
create the app config dir.

In order to avoid such problems, ensure the user config dir exists
before trying to create app config dir.