Re #170: Try to optimise large image paste
Up to 256px is the max likely icon size, so check all pixels
(we'll stop at the first non-opaque one, so icons will return quickly)

Over that, check a staggered grid at 7px increments.
The prime number and shifting offset *should* mean that we're
not going to keep hitting any transparent grids and allows us
to do 1/49th of the processing!

The lag is still in the initial "fetch from clipboard", though,
and is beyond our control.