Re #396: Implement "refresh" by keyboard shortcut
Take the quick and ugly way of clearing the model and loading old
tweets.

Works on home, mentions and likes.

NOTE: A simple cap has been added to prevent repeated refreshes
because a) we already poll every 2 minutes and b) it wastes both
the user's 15/75 requests per 15 minutes *and* the global 100K/day
cap that Twitter added in June 2019.

Fixes #396